# docs/api/swagger.yaml
openapi: 3.0.0
paths:
  /users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '409':
          description: ''
        '500':
          description: ''
      tags: &a1
        - users
    get:
      operationId: UsersController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: username
          required: false
          in: query
          schema:
            minLength: 1
            maxLength: 25
            example: Alice
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersPaginatedResponseDto'
        '500':
          description: ''
      tags: *a1
  /users/provider/{provider}/{providerUserId}:
    get:
      operationId: UsersController_getByIdentity
      parameters:
        - name: provider
          required: true
          in: path
          schema:
            type: string
        - name: providerUserId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: ''
        '500':
          description: ''
      tags: *a1
  /users/{id}:
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: ''
        '409':
          description: ''
        '500':
          description: ''
      tags: *a1
  /users/{id}/confirm:
    post:
      operationId: UsersController_confirm
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmUserResponseDto'
        '500':
          description: ''
      tags: *a1
info:
  title: HaxFootball API
  description: ''
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateUserDto:
      type: object
      properties:
        provider:
          type: string
          example: discord
        providerUserId:
          type: string
          example: '123456789012345678'
        username:
          type: string
          minLength: 1
          maxLength: 25
          example: Alice
        password:
          type: string
          nullable: true
          minLength: 4
          maxLength: 20
        role:
          type: string
          enum: &a2
            - admin
            - mod
            - default
          example: default
      required:
        - provider
        - providerUserId
        - username
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        tenant:
          type: string
        provider:
          type: string
        provider_user_id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: *a2
        created_at:
          type: string
          format: date-time
      required:
        - id
        - tenant
        - provider
        - provider_user_id
        - username
        - role
        - created_at
    PageInfoDto:
      type: object
      properties:
        page:
          type: number
        page_size:
          type: number
        has_next_page:
          type: boolean
      required:
        - page
        - page_size
        - has_next_page
    UsersPaginatedResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseDto'
        page_info:
          $ref: '#/components/schemas/PageInfoDto'
      required:
        - items
        - page_info
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 25
          example: Alice
        password:
          type: string
          nullable: true
          minLength: 4
          maxLength: 20
        role:
          type: string
          enum: *a2
          example: mod
    ConfirmUserDto:
      type: object
      properties:
        password:
          type: string
          minLength: 4
          maxLength: 20
          example: secret-password
      required:
        - password
    ConfirmUserResponseDto:
      type: object
      properties:
        is_correct:
          type: boolean
        provider_user_id:
          type: string
          nullable: true
      required:
        - is_correct
        - provider_user_id
security:
  - bearer: []